Update __cts__.xml
So I have proposed a change here, but I'm not sure this is technically correct.  I was wondering what I should do about the HTML character entity here for the equals sign, which wasn't correct anyway since it should have been &#61.  I have copied in in the proper TLG title here: Apotelesmatica (= Tetrabiblos) since I'm not sure how to represent the = sign in the title.