Added in author. Standardized metadata. Changed file ID.
Also: Should we change the title of this work as it does not reflect the reality of the file? Only volume 1 of Dindorf's edition
of this work seems to be contained in this file, but I couldn't quite figure out how many books of the full work were within this file.