Added xml:lang tags where they were missing
These tags were automatically taken from the URNs of the edition or translation. So if the last 4 letters of the URN are grc1, the work edition was assigned and xml:lang attribute of 'grc'. If 'ger1', then of 'deu'. Etc. This automatic insertion could have hit problems if the URNs were produced strangely. But that does not seem to be a problem.